1_messaging_sqs_lambda_rds/requirements.txt
mysql-connector-python
boto3
python-dotenv